functions/requirements.txt
firebase-functions
firebase-admin
vertexai
numpy
google-genai